add SERCQ value to channelType enumeration
