+++ packages/pn-personafisica-webapp/src/models/contacts.ts
@@ -3,6 +3,7 @@ export enum ChannelType {
   EMAIL = 'EMAIL',
   SMS = 'SMS',
   IOMSG = 'APPIO',
+  SERCQ = 'SERCQ',
 }
 
 export enum IOAllowedValues {
@@ -23,4 +24,19 @@ export interface DigitalAddress {
   value: string;
   pecValid?: boolean;
   codeValid?: boolean;
+}
+
+export interface DeleteDigitalAddressParams {
+  addressType: AddressType;
+  senderId: string;
+  channelType: ChannelType;
+}
+
+export interface SaveDigitalAddressParams {
+  addressType: AddressType;
+  senderId: string;
+  senderName?: string;
+  channelType: ChannelType;
+  value: string;
+  code?: string;
 }
