+++ packages/pn-personagiuridica-webapp/src/components/Contacts/SpecialContacts.tsx
@@ -34,14 +34,18 @@ import {
 } from '@pagopa-pn/pn-commons';
 import { ButtonNaked } from '@pagopa/mui-italia';
 
-import { AddressType, ChannelType, DigitalAddress } from '../../models/contacts';
+import {
+  AddressType,
+  ChannelType,
+  DigitalAddress,
+  SaveDigitalAddressParams,
+} from '../../models/contacts';
 import { Party } from '../../models/party';
 import {
   CONTACT_ACTIONS,
   createOrUpdateAddress,
   getAllActivatedParties,
 } from '../../redux/contact/actions';
-import { SaveDigitalAddressParams } from '../../redux/contact/types';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
 import { RootState } from '../../redux/store';
 import {
