+++ packages/pn-personagiuridica-webapp/src/models/contacts.ts
@@ -2,6 +2,7 @@ export enum ChannelType {
   PEC = 'PEC',
   EMAIL = 'EMAIL',
   SMS = 'SMS',
+  SERCQ = 'SERCQ',
 }
 
 export enum AddressType {
@@ -17,4 +18,19 @@ export interface DigitalAddress {
   value: string;
   pecValid?: boolean;
   codeValid?: boolean;
+}
+
+export interface DeleteDigitalAddressParams {
+  addressType: AddressType;
+  senderId: string;
+  channelType: ChannelType;
+}
+
+export interface SaveDigitalAddressParams {
+  addressType: AddressType;
+  senderId: string;
+  senderName?: string;
+  channelType: ChannelType;
+  value: string;
+  code?: string;
 }
