+++ packages/pn-personagiuridica-webapp/src/redux/contact/actions.ts
@@ -2,11 +2,15 @@ import { parseError } from '@pagopa-pn/pn-commons';
 import { createAsyncThunk } from '@reduxjs/toolkit';
 
 import { apiClient } from '../../api/apiClients';
-import { AddressesApiFactory } from '../../generated-client/digital-addresses';
+import { AddressesApiFactory, BffChannelType } from '../../generated-client/digital-addresses';
 import { InfoRecipientApiFactory } from '../../generated-client/recipient-info';
-import { AddressType, DigitalAddress } from '../../models/contacts';
+import {
+  AddressType,
+  DeleteDigitalAddressParams,
+  DigitalAddress,
+  SaveDigitalAddressParams,
+} from '../../models/contacts';
 import { FilterPartiesParams, Party } from '../../models/party';
-import { DeleteDigitalAddressParams, SaveDigitalAddressParams } from './types';
 
 export enum CONTACT_ACTIONS {
   GET_DIGITAL_ADDRESSES = 'getDigitalAddresses',
@@ -42,7 +46,8 @@ export const createOrUpdateAddress = createAsyncThunk<
       const response = await digitalAddressesFactory.createOrUpdateAddressV1(
         params.addressType,
         params.senderId,
-        params.channelType,
+        // TODO: rimuovere l'as appena si ha il bff integrato
+        params.channelType as BffChannelType,
         { value: params.value, verificationCode: params.code }
       );
 
@@ -92,7 +97,8 @@ export const deleteAddress = createAsyncThunk<void, DeleteDigitalAddressParams>(
       const response = await digitalAddressesFactory.deleteAddressV1(
         params.addressType,
         params.senderId,
-        params.channelType
+        // TODO: rimuovere l'as appena si ha il bff integrato
+        params.channelType as BffChannelType
       );
 
       return response.data;
